Add unit tests improve code

+++ TheShop/Contracts/IShopService.cs
@@ -9,7 +9,7 @@ namespace TheShop.Contracts
 {
     public interface IShopService
     {
-        Article OrderArticle(int id, int maxExpectedPrice);
+        Article OrderArticle(int id, decimal maxExpectedPrice);
         void SellArticle(int buyerId, Article article);
         void DispalyArticle(Article article);
     }
